feat(mfi-v2-ui): Point page restructuring
