+++ apps/marginfi-v2-ui/src/components/desktop/Points/index.ts
@@ -1,0 +1,6 @@
+export * from "./PointsOverview";
+export * from "./PointsLeaderBoard";
+export * from "./PointsSignIn";
+export * from "./PointsSignUp";
+export * from "./PointsCheckingUser";
+export * from "./PointsConnectWallet";
